Copy-edit, change header styling, fix scrollbar
Make various changes to wording for clarity. Remove link to spoiler
filled interview. Change header styling so that outside of bar is
outset and colored, and left and right links have rounded edges.
Fix scrollbar gutter so the pages stay in the same place whether
it is there or not.